Changing fn signature inconsistency

+++ Medicaid.h
@@ -15,9 +15,9 @@ typedef struct {
 } MedicaidPacket;
 
 typedef struct {
-    uint32_t current_queue_len;
-    uint32_t total_dropped;
-    uint32_t avg_latency_ms; // Running average
+    uint32_t current_queue_len; // Backlog of cases to process
+    uint32_t total_dropped;     // Total number of cases dropped due to expiration
+    uint32_t avg_latency_ms;    // Running average - represents average medicaid processing time
 } ServerReport; // 12 Bytes
 
 
@@ -25,7 +25,7 @@ int serialize_packet(MedicaidPacket* packet, uint8_t* buffer, size_t buffer_size
 int deserialize_packet(MedicaidPacket* packet, uint8_t* buffer, size_t buffer_size);
 uint64_t my_htonll(uint64_t val);
 uint64_t my_ntohll(uint64_t val);
-MedicaidPacket create_medicaid_packet(uint32_t id, uint32_t zone, uint32_t timestamp_ms, uint8_t complexity, uint8_t manual);
+MedicaidPacket create_medicaid_packet(uint32_t id, uint32_t zone, uint64_t timestamp_ms, uint8_t complexity, uint8_t manual);
 int send_medicaid_packet(MedicaidPacket* packet, int sock_fd);
 int recv_medicaid_packet(MedicaidPacket* packet, int sock_fd);
 int init_packet_struct(MedicaidPacket* packet);
